requirements.txt
streamlit
Pillow
scikit-image
streamlit-image-comparison
